refactor: split processor_hc into smaller modules
